updated getCreateTableQuery method. Comments added. Moved base class to framework pkg.
