+++ library/src/main/java/com/blueshift/httpmanager/request_queue/RequestQueue.java
@@ -10,7 +10,6 @@ import com.blueshift.Blueshift;
 import com.blueshift.httpmanager.HTTPManager;
 import com.blueshift.httpmanager.Request;
 import com.blueshift.httpmanager.Response;
-import com.blueshift.httpmanager.request_queue.db.RequestQueueTable;
 import com.blueshift.model.Configuration;
 
 /**
